requirements.txt
# --- Core DL Framework ---
torch==2.9.0
torchvision==0.24.0
numpy==2.2.5
pandas==2.3.3

# --- CV & Pathology Specific ---
openslide-python==1.4.2
pillow==11.3.0
opencv-python==4.12.0
webdataset==1.0.2

# --- Model Architectures & Utilities ---
timm==1.0.21
transformers==4.57.1
einops==0.8.1
safetensors==0.6.2

# --- Data Processing & Misc ---
tqdm==4.67.1
pyyaml==6.0.3
requests==2.32.5
huggingface-hub==0.36.0
networkx==3.3
typing-extensions==4.15.0